Move command parsing to own file
